==> notebook/01_dataset_preprocessing.ipynb <==
import pandas as pd
import os

df_data = pd.read_csv("../input/patient_priority.csv")
df_data.head(5)

#### Basic description of the dataset

df_data.shape

df_data.describe()

#### Check for null values

df_data.info()

df_data.isnull().sum()

df_data = df_data.dropna() # eliminar la fila que tiene valores vacios en genero y triage

#### Check for duplicate rows

duplicate = df_data.duplicated()
print(f"Total duplicate rows: {duplicate.sum()}")

#### View data types

df_data.dtypes

cols_to_int = ['age', 'gender', 'chest pain type', 'exercise angina', 'hypertension', 'heart_disease']

# Convertir las columnas a tipo int
df_data[cols_to_int] = df_data[cols_to_int].astype(int)

df_data.dtypes

df_data["Residence_type"].value_counts()

df_data["smoking_status"].value_counts()

#### Change column values ​​to numeric values

df_encoded = pd.get_dummies(df_data, columns=['Residence_type', 'smoking_status'])

# Convertir solo las columnas generadas por one-hot encoding a int
one_hot_columns = [col for col in df_encoded.columns if col.startswith('Residence_type_') or col.startswith('smoking_status_')]
df_encoded[one_hot_columns] = df_encoded[one_hot_columns].astype(int)

#### Undersampling to balance the dataset

df_encoded["triage"].value_counts()

min_size = df_encoded["triage"].value_counts().min()

undersampled_df = (
    df_encoded.groupby("triage")
    .apply(lambda x: x.sample(min_size, random_state=42))  # Selecciona aleatoriamente
    .reset_index(drop=True)
)

undersampled_df["triage"].value_counts()

#df_encoded = undersampled_df

#### Save dataset and response column

df_labels = df_encoded[["Unnamed: 0", "triage"]]
df_labels = df_labels.rename(columns={"Unnamed: 0": "id"})

df_encoded = df_encoded.drop(["Unnamed: 0", "triage"], axis=1)

os.makedirs("../results/preprocessing/", exist_ok=True)
df_labels.to_csv("../results/preprocessing/labels.csv", index=False)
df_encoded.to_csv("../results/preprocessing/01_dataset_preprocessing.csv", index=False)
