==> notebook/03_evaluation_outliers.ipynb <==
#### Outlier detection

import sys
import pandas as pd
sys.path.insert(0, "../")

from src.preprocessing.outlier_detection import OutlierDetection

df_data = pd.read_csv('../results/preprocessing/02_dataset_preprocessing.csv')
detector = OutlierDetection(df_data=df_data)

##### Exploring outliers with local method

outliers_iqr = detector.detect_outliers_iqr()
outliers_iqr

#### Outlier detection with global method

result_if = detector.filter_isolation_forest()
outliers_if = result_if[result_if['outlier'] == -1] # predicción de outliers (-1 para outliers, 1 para normales)
outliers_if

result_if['outlier'].value_counts()
