==> notebook/05_scaler_data.ipynb <==
#### Scale numeric descriptors

import sys
import pandas as pd
sys.path.insert(0, "../")

from src.preprocessing.scaler import DataScaler

df_data = pd.read_csv("../results/preprocessing/02_dataset_preprocessing.csv")
scaler = DataScaler(data=df_data)

#### Separating non-binary data for scaling
Scaling binary data can alter its meaning and these are already in the appropriate scale (0 and 1). The same occurs with the column `chest pain type` with value in range 0 to 4.

df_binary, df_non_binary = scaler.separating_data()

#### Scaling the data with Robust Scaler
Robust Scaler is used because the data has some outlier examples and it is a scaler that is not very sensitive to outliers

df_scaled, scaler_instance = scaler.scale_data(df_non_binary=df_non_binary)
df_scaled.head()

#### Save data

scaler.save_results(df_binary=df_binary,
                    df_scaled=df_scaled,
                    scaler_instance=scaler_instance,
                    output_csv="../results/preprocessing/05_dataset_scaled_robust.csv",
                    output_scaler="../results/preprocessing/scaler_robust.joblib"
)
