==> notebook/07_apply_clustering.ipynb <==
#### Applies clustering algorithms

import sys
import pandas as pd
sys.path.insert(0, "../")

from src.clustering.apply_clustering import ClusteringAnalysis

df_data = pd.read_csv("../results/preprocessing/05_dataset_scaled_robust.csv")
df_labels = pd.read_csv("../results/preprocessing/labels.csv")

clustering = ClusteringAnalysis(df_data=df_data,
                                df_labels=df_labels)
clustering.run_all_algorithms()
clustering.save_results(export_path="../results/clustering")
