==> notebook/12_apply_regression_model.ipynb <==
import sys
import pandas as pd
sys.path.insert(0, "../")

from sklearn.model_selection import train_test_split
from src.regression.regression_models import RegressionModel

### Load generated dataset

df_data = pd.read_csv("../results/preprocessing/05_dataset_scaled_robust.csv")
df_labels_clustering = pd.read_csv("../results/clustering/labels.csv")

df_data = df_data.assign(label=df_labels_clustering["minibatch_k_means_k_4"]) # algoritmo con mejor desempeño
responses = df_data.pop("blood pressure").values # modelos de regresion para presion arterial

train_data, validation_data, train_response, validation_response = train_test_split(df_data, responses, random_state=42, test_size=.30)

### Training model

regressor = RegressionModel(
    train_values=train_data,
    test_values=validation_data,
    train_response=train_response,
    test_response=validation_response,
    folder_export="../results/regression"
)

df_summary = regressor.apply_regression()

df_summary
