==> notebook/14_xai_shap_analysis.ipynb <==
import sys
import pandas as pd
sys.path.insert(0, "../")

from src.xai.apply_shap import SHAPAnalyzer

#### Load dataset

df_data = pd.read_csv("../results/classification/10_classification_data.csv")
data_values = df_data.drop(columns=['label'])
responses = df_data['label']

#### SHAP analyzer by model

# Analizar con Decision Tree
analyzer_dt = SHAPAnalyzer("decision_tree", data_values, responses)
analyzer_dt.analyze(save_dir="../results/xai/decision_tree/")

# Analizar con Random Forest
analyzer_rf = SHAPAnalyzer("random_forest", data_values, responses)
analyzer_rf.analyze(save_dir="../results/xai/random_forest/")

# Analizar con Gradient Boosting
analyzer_rf = SHAPAnalyzer("gradient_boosting", data_values, responses)
analyzer_rf.analyze(save_dir="../results/xai/gradient_boosting/")

# Analizar con adaboost
analyzer_rf = SHAPAnalyzer("adaboost", data_values, responses)
analyzer_rf.analyze(save_dir="../results/xai/adaboost/")

# Analizar con extra_trees
analyzer_rf = SHAPAnalyzer("extra_trees", data_values, responses)
analyzer_rf.analyze(save_dir="../results/xai/extra_trees/")

# Analizar con KNN
analyzer_rf = SHAPAnalyzer("knn", data_values, responses)
analyzer_rf.analyze(save_dir="../results/xai/knn/")

# Analizar con SVC
analyzer_rf = SHAPAnalyzer("svc", data_values, responses)
analyzer_rf.analyze(save_dir="../results/xai/svc/")
