==> notebook/15_xai_counterfactuals_exploring.ipynb <==
import sys
import pandas as pd
sys.path.insert(0, "../")

from src.xai.apply_counterfactuals import CounterfactualExplainer

#### Load dataset

df_data = pd.read_csv("../results/classification/10_classification_data.csv")
data_values = df_data.drop(columns=['label'])
responses= df_data['label']

#### Generate counterfactuals by class

explainer = CounterfactualExplainer("../results/classification/10_classification_data.csv")
query_instance = explainer.train_data.iloc[[0]] # Seleccionar una instancia de consulta

counterfactuals_list = explainer.generate_and_show_counterfactuals(query_instance, total_cfs=19)
